requirements.txt
streamlit
google-generativeai
tensorflow-cpu
numpy
requests
Pillow